pyproject.toml
[tool.poetry]
name = "aggresive-moderator-bot"
version = "0.1.1"
description = ""
authors = ["Jon Besga <jonan.bsg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.7"
openai = "^1.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
